MOT:2119 OSP 1.0 doesn't report SIGNIFICANT_MOTION event
MOT:2116 OSP 1.0 doesn't report Step Detect and Count

This Patchset contains below fixes :
1. Step detector : Created IIO device with buffer enabled to report
step detection to Android. Also changed sensor name in N7.config
as names were same for step counter and step detector sensor.
2. Significant Motion : Enabled IIO buffer for significant motion sensor
to report significant motion event.
3. Removed Pressure sensor IIO device in kernel since we are not using it.
Changed init.grouper.rc file as IIO device numbering will change after removal
of pressure sensor device.
4. Add flush failure condition for one shot and wake up sensors.

+++ ospd/OSPDaemon.h
@@ -120,7 +120,7 @@ static inline void conv2OSP(ASensorType_t sensor,
 		break;
 	case SENSOR_SIGNIFICANT_MOTION:
 	case SENSOR_STEP_DETECTOR:
-		d->booldata.data = val[0];
+		d->booldata.data = (uint8_t)val[0];
 		d->booldata.TimeStamp = ts;
 		break;
 	case SENSOR_STEP_COUNTER:
